Today marks the beginning of the implementation of one of the last big additions to PeasantRPG's v1.0.0 Beta update: A modular world system. This change is going to be HUGE, and will require a complete rewrite of many different parts of the game. So far, almost nothing in the game works. The movement_system() function in the world.py file is the actual game loop for PeasantRPG, and it isn't functional yet. HOWEVER, I've already made decent progress and the system will be usuable hopefully later today.
The main goal of this system to allow each individual world tile to be completely unique. No longer will the world of Harconia just be a large square with different rectangular regions in it. After this update, the world will be massive and complete, and will roughly be the actual shape of Harconia.

Unlike before, each new tile can have several interesting properties:
	> A custom level requirement
	> A specific level for all monsters located there
	> A custom description
	> A different region, meaning that each region can be as big or as small or whatever shape I want it to be
	> Support for more than one town per tile, and support for going UP and DOWN, instead of just NSEW
	> Support for different exits and enterences. This means that the Aether doesn't have to be programmed
as a weird layer above the existing overworld - instead, it will just be an extension of the normal overworld.

This new Tile system marks a HUGE step forward towards making this game fun and exciting, and there's no limit to what I can do with it.