Fixed up a lot of code. The game no longer crashes when you attempt to load a file. Added a date-time feature as well as numerous additional logging functions to the loading/unloading functions. Fixed some PCUs and monsters having their HP listed as MP and vice versa. Added a random_encounter.wav file that could potentially be used when encountering an enemy on the overworld. Fixed the "c_spell" attribute crashing the game if a spell is used before saving.
Reworked the UI when choosing a class: The game now lists a 6-8 point list of unique characteristics for each class, sucha as Mage's ability to learn all spells and Paladin's increase Wisdom->Healing bonus. It also says what damage type standard attacks deal, what type of damage their weapon increases, what their main attribute is, and what their stat distribution is.

Reworked the Movement HUD: Instead of listing all available commands every time you move, the game now only says "Input command". Typing "Help" will display a list of several commands you can use. This is really cool because it means that I can add dozens of different commands without having to constantly list all of them every time the player moves. This is also similar to how things work in most adventure games.

Added an "Autosave" function in the form of a "verbose" argument in the save_game() function. Calling this function with verbose set to False will cause no text to be printed, no inputs to be requested, and no smart_sleep() functions to be called. Currently this is being used to immediately save the game when the player creates a new save file, but modders or people using this as a game engine could potentially do much more with it. This, by the way, is just one of several reworks that are coming to the save feature.