Nerfed the design for Skill Shot and Canonize. They were originally supposed to be the Mage/Paladin class's ultimates, but I came up with a much better idea for ultimates.
Fixed up the settings.cfg file: Removed the save file name settings (their functionality was removed several weeks ago because they were super buggy and didn't serve much of a purpose anyway. This is just clearing out the last remnants of it). Added in a new setting that allows people to change the size of the divider line (the '-'*25).

Split the PeasantsAscension.py file into three separate files: PeasantsAscension.py, which remains the __main__ file; save_load.py, which is used to save/load the game and change the settings; and title_screen.py which is used to display the title screen and handle printing the options on the title menu. This is also accompionied by splitting the save_game() function into save_game() and serialize_all() funcitons, and the load_game() (renamed from check_save()) into load_game() and deserialize_all(). The reason I did this was because the PeasantsAscenion.py file was getting too big, and some upcoming reworks to the saving system made this a good solution.