Fixed the setup.py file by having it add the new folders to __path__. Moved a large chunk of the movement_system() function to a new "move_command()" function. Removed the find_tile_with_coords() function, because due to oversights it didn't work and I didn't know how to fix it. I managed to cut the magic_command() function down from 40 lines to only 9 using the Unit.choose_target() method I made a while ago. Removed "c" and "cancel" from being valid options to exit menus in several locations.
Added a new "debug_command()" function that can be accessed on the overworld by typing in "debug-menu". This command is VERY powerful, allowing the player to type in pure python to manipulate in-game variables. It's also very dangerous, as if the player does not know what they're doing, they can very easily make the game unplayable. As such, it's not listed in the help menu and is only to be used for testing purposes. Removed the "give me the gold" cheat, because there's no point to it anymore thanks to the debug menu.