Nerfed Great Cleave to be a 1.25x crit followed by a [150 + Strength]% Crit. Reworked Tip the Scales to scale off max HP, instead of being a flat bonus. Reworked Unholy Binds to have a [10 + Wisdom]% chance of instantly killing the target if the target is already Darkness. Changed Judgement to synergize with Unholy Binds, lowering the insta-death timer. Also made it not work on bosses. Reworked infusion to let the player choose a random element in exchange for a bigger damage buff.
Reworked the way that Conversations are handled. Instead of being a long list of strings, each of which has to be manually set to 79 characters long, dialogue is now just one big multiline string. This multiline string is automatically chopped into a list with 79-length strings using a new Conversation.chop() method. Added an "active" attribute to the NPC class. Removed the end_dialogue and desc attributes from the Quest class - the former is now handled using conversations, and the latter is replaced with the original dialogue. This is a WIP, a lot of things are broken right now.