Fully named all 12 provinces, and added in support for two new biomes (dungeon/castle, and underworld). Renamed Sadora (which I thought sounded to much like another province, Peladora) to "Deltora", after the popular Deltora Quest books. Also renamed Peladora to "Pelamora", because apparently Peladora is the Spanish word for grill or something like that and I don't want people to think there's some sort of symbolism there when there isn't.
Cleaned up the watermelon_store() function to be more in-line with the current style I'm using for the rest of the game. Added an easter egg for if the player somehow does manage to get enough gold to buy the uber-expensive items at that store.

Cleaned up the coordinate system, because I realized that coord_change was being calculated in two different places when only one was needed. This might have been the cause of the desynced coordinates issue, so hopefully that bug is fixed (I don't know when/how/why it happens so I can't really test it).

Also fixed a few minor issues.