Began work on revamping the settings system: Settings can now be viewed and changed in-game, and apply immediately upon changing. Currently this only works for changing music and sound volume, but it will be trivial to implement it for divider size and blip toggling when I have some more free time.
Added a new unequip_item() function, which does exactly what you think it does. To accompiany this, I've added 4 new "None" items, similar to the Fists. These items cannot be unequipped and are simply there to prevent the player having nothing equipped. When equipping an item over it, the nothing item is simply deleted from existance, and added back if the new item is unequipped.

Lowered the volume of ALL Music and Sound FX by 10-25 dBs, because I've gotten complaints from friends that they're way too loud and I definitely agree.

Removed the __str__() method from the Item, Spell, and NPC classes, to improve readability. Improved the auto-debugger a bit more (it can now buy from stores and enter unlocked houses, and will not choose to view settings or exit the game). Removed some grammar checking stuff that made the code look ugly. This means that the grammar will be wonky in some areas but no other RPG I've seen had grammar checking so I doubt people will care. Removed the old graveyard music (it was pretty bad) and replaced it with the old cave music. Renamed the dungeon biome to "castle", and added in Ben Landis's castle music for it. The item_pickup sound effect now stops playing in some important places.

There's still that bug with the coordinates... When I release v1.0.0 I will probably set a $50 bounty or something if someone can figure out what the deal is. It's kind of frustrating how much more work there is to do on this game, but it's fun so no big deal.