Fixed the auto-debugger not being able to put skill points into FOR, CON, INT, and DEX. Fixed the auto-debugger not being able to access the inventory. Added more friend names and removed ones that are likely to be peoples actual names.
Removed Skill Shot and designed a new skill, "Skill Syphon". This ability is much cooler IMO, and if it's too OP it'd be easier to adjust than the previous one. Probably need to do the same for Ranger's Unstable Footing, because that one doesn't make very good use of attributes and overall would probably be really hard to make useful.