Added Cells, which contain tiles, and Provinces, which contain cells. Cells are the small points of interest that I mentioned a few months ago, such as towns and bandit hideouts. Provinces are a set of 12 areas that make up the entirety of the world map, each with their own set of cells. Each province can only be visited once the proper story conditions have been met. Both provinces and cells have their own dedicated classes to make things simpler. Biomes are now cell-specific instead of tile-specific. Province names are now province-specific instead of tile-specific.
Added the "Fast Travel Atlus", which allows the player to teleport to any available province/cell. This is currently unobtainable but will be given to the player as soon as they obtain Solou as a party member.
Added a new "gamestate" value in the party_info dictionary that can be used to check whether the player in a battle, town, or the overworld.
Bosses can now have lackies, which are additional monsters that fight on their side.
Removed the Divining Rod and the Watermelon Store
the iSound has been renamed to the Portable Music Box and can now be bought at any general store.
Renamed Simphet to Parriway.
Cleaned up the tools menu.