Programmed the portable musicbox! I'm really proud of this, I had to make a file explorer, use multiprocessing, and create like a billion methods to do this and it all turned out really well.
Created a dedicated play_music() function. This is a one-line substitute for the previous three lines needed to play music, but it also lets me conditionally not play music if the musicbox is running.
Fixed main.s_input() not being able to accept nothing as an argument.
Removed the files I used to test my musicbox concept.