Fixed a few bugs with the musicbox - it can now play mp3 files, the music now stops if you delete the musicbox, and haivng multiple musicboxes and turning off the "wrong" one no longer crashes the game.
Items that can be used in battle now use the gamestate key in the party info dictionary instead of an is_battle argument. Renamed Ran'af to "Storm", who is a character from the RP. All use_item() methods now have a user argument to simplify things. Rewrote some of the code related to the inventory and selling items.