Expanded the list of characters that the auto-tester can choose from. This means it could now do things like enter general stores and view quest information, which lead me to several game-crashing bugs. Fixed!
This also led me to find some other issues, which I haven't fixed yet. The big one is that entering 0 on an index-based menu, like the file select screen, will cause it to choose the last element of the list (this is completely logical, because indexes are zero-based, meaning that the chosen number needs one subtracted from it to equal the true index, and -1 maps to the last element in an iterable in python). Just an oversight, I'll fix it later.

I also REALLY need to rewrite several functions, namely some of the town functions and the view_quests() function. Those are messy!