Finished the alchemy lab! Now I just have to program the potions.
Removed the separate buy and sell values for items. Instead, items now have a singular "value" attribute. Selling an item gives you gold equal to 20% of the value. This is to help prepare for the future implementation of the CHARISMA attribute, which will be the main attribute of Bards (new class!).

Weapons and armor now cost way more.