The Bard will now have 6 abilities instead of 10, because 10 is honestly too much work for me. Bards now have terrible magical attack instead of stong magic attack, and their attacks no longer miss. This is because I want them to be more focused on supporting and combo-ing their abilities than just attacking things to death. This means that Bards now have less than the standard 16 stat spread per level, which is compensated for with the additional 2 abilities. Designed the Stubborn Boar ability.
Removed the text_scroll and TownClass modules - the former is now built into the main module, and the latter is now combined with the town module. All towns now have their own Class. I'm planning on reworking the class structure for Towns so that I can have camps solely for speaking to NPCs, without general stores or inns.