Cleaned up the tiles.py file. All cells now have their own class, and all tiles are now created inside the .__init__() for the cell they belong to. Just looks much cleaner, no real changes.
The main reason I did this now instead of later is that I have a neat plan that I'm going to execute inside this file, and it'll make it much easier if the file is cleaned up like this. I'm planning on adding a few dungeon-crawling elements, so get excited for that.