Revamped input!
Several inputs no longer require the user to press enter, this is done through SingleCharInput().
Others still need an enter press so input can be multiple characters long, this is done through MultiCharInput().
Some need a key press in some situations, and an enter press in others, this is done through FlexibleInput().
And finally a few inputs don't actually need to know what key was pressed, this is done through PressAnyKeyToContinue().

Overall the game is a lot quicker to play now!